requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
langchain-community==0.0.38
langchain-chroma==0.1.2
langchain-ollama==0.1.0
chromadb==0.4.18
ollama==0.1.7
python-multipart==0.0.6
